Update Game.java
Created a main menu for the game with a function to start a new game and load a saved game. Implemented serializable into Game.java so that the game can be saved. Made a new function that the user can save the game and continue the game. Also added an auto save when the user exited the game.